+++ src/index.ts
@@ -2,5 +2,6 @@ import Button from "./components/Button";
 import TextInput from "./components/TextInput";
 import ToggleButton from "./components/ToggleButton";
 import MultilineInput from "./components/MultilineInput";
+import SelectInput from "./components/Select";
 
-export { Button, TextInput, ToggleButton, MultilineInput };
+export { Button, TextInput, ToggleButton, MultilineInput, SelectInput };